+++ lib/searchkick/index.rb
@@ -17,7 +17,7 @@ module Searchkick
     end
 
     def delete
-      if !Searchkick.server_below?("6.0.0") && alias_exists?
+      if alias_exists?
         # can't call delete directly on aliases in ES 6
         indices = client.indices.get_alias(name: name).keys
         client.indices.delete index: indices
@@ -68,7 +68,7 @@ module Searchkick
           }
         )
 
-      response["hits"]["total"]
+      Searchkick::Results.new(nil, response).total_count
     end
 
     def promote(new_name, update_refresh_interval: false)