+++ test/aggs_test.rb
@@ -20,8 +20,7 @@ class AggsTest < Minitest::Test
   end
 
   def test_order
-    order_key = Searchkick.server_below?("6.0") ? "_term" : "_key"
-    agg = Product.search("Product", aggs: {color: {order: {order_key => "desc"}}}).aggs["color"]
+    agg = Product.search("Product", aggs: {color: {order: {_key: "desc"}}}).aggs["color"]
     assert_equal %w(red green blue), agg["buckets"].map { |b| b["key"] }
   end
 
@@ -37,8 +36,7 @@ class AggsTest < Minitest::Test
 
   def test_script
     source = "'Color: ' + _value"
-    script = Searchkick.server_below?("5.6") ? {inline: source} : {source: source}
-    agg = Product.search("Product", aggs: {color: {script: script}}).aggs["color"]
+    agg = Product.search("Product", aggs: {color: {script: {source: source}}}).aggs["color"]
     assert_equal ({"Color: blue" => 1, "Color: green" => 1, "Color: red" => 1}), buckets_as_hash(agg)
   end