+++ test/errors_test.rb
@@ -4,13 +4,15 @@ class ErrorsTest < Minitest::Test
   def test_bulk_import_raises_error
     valid_dog = Product.create(name: "2016-01-02")
     invalid_dog = Product.create(name: "Ol' One-Leg")
-    index = Searchkick::Index.new "dogs", mappings: {
-      dog: {
-        properties: {
-          name: {type: "date"}
-        }
+    mapping = {
+      properties: {
+        name: {type: "date"}
       }
     }
+    mapping = {product: mapping} if Searchkick.server_below?("7.0.0")
+    index = Searchkick::Index.new "dogs", mappings: mapping
+    index.delete if index.exists?
+    index.create_index
     index.store valid_dog
     assert_raises(Searchkick::ImportError) do
       index.bulk_index [valid_dog, invalid_dog]