+++ test/models/store.rb
@@ -1,14 +1,15 @@
 class Store
+  mappings = {
+    properties: {
+      name: {type: "keyword"}
+    }
+  }
+  mappings = {store: mappings} if Searchkick.server_below?("7.0.0")
+
   searchkick \
     routing: true,
     merge_mappings: true,
-    mappings: {
-      store: {
-        properties: {
-          name: {type: "keyword"}
-        }
-      }
-    }
+    mappings: mappings
 
   def search_document_id
     id