+++ test/multi_indices_test.rb
@@ -7,16 +7,44 @@ class MultiIndicesTest < Minitest::Test
     assert_search_multi "product", ["Product A", "Product B"]
   end
 
-  def test_where
-    store [{name: "Product A", color: "red"}, {name: "Product B", color: "blue"}]
-    store_names ["Product C"], Speaker
-    assert_search_multi "product", ["Product A", "Product C"], where: {_or: [{_type: "product", color: "red"}, {_type: "speaker"}]}
+  def test_index_name
+    store_names ["Product A"]
+    assert_equal ["Product A"], Product.search("product", index_name: Product.searchkick_index.name).map(&:name)
+    assert_equal ["Product A"], Product.search("product", index_name: Product).map(&:name)
+    assert_equal [], Product.search("product", index_name: Speaker.searchkick_index.name, conversions: false).map(&:name)
+  end
+
+  def test_models_and_index_name
+    store_names ["Product A"]
+    store_names ["Product B"], Speaker
+    assert_equal ["Product A"], Searchkick.search("product", models: [Product, Store], index_name: Product.searchkick_index.name).map(&:name)
+    error = assert_raises(Searchkick::Error) do
+      Searchkick.search("product", models: [Product, Store], index_name: Speaker.searchkick_index.name).map(&:name)
+    end
+    assert_includes error.message, "Unknown model"
+    # legacy
+    assert_equal ["Product A"], Searchkick.search("product", index_name: [Product, Store]).map(&:name)
+  end
+
+  def test_model_with_another_model
+    error = assert_raises(ArgumentError) do
+      Product.search(models: [Store])
+    end
+    assert_includes error.message, "Use Searchkick.search"
+  end
+
+  def test_model_with_another_model_in_index_name
+    error = assert_raises(ArgumentError) do
+      # legacy protection
+      Product.search(index_name: [Store, "another"])
+    end
+    assert_includes error.message, "Use Searchkick.search"
   end
 
   private
 
   def assert_search_multi(term, expected, options = {})
-    options[:index_name] = [Product, Speaker]
+    options[:models] = [Product, Speaker]
     options[:fields] = [:name]
     assert_search(term, expected, options, Searchkick)
   end