+++ test/query_test.rb
@@ -4,16 +4,9 @@ class QueryTest < Minitest::Test
   def test_basic
     store_names ["Milk", "Apple"]
     query = Product.search("milk", execute: false)
-    # query.body = {query: {match_all: {}}}
-    # query.body = {query: {match: {name: "Apple"}}}
     query.body[:query] = {match_all: {}}
     assert_equal ["Apple", "Milk"], query.map(&:name).sort
     assert_equal ["Apple", "Milk"], query.execute.map(&:name).sort
-  end
-
-  def test_with_effective_min_score
-    store_names ["Milk", "Milk2"]
-    assert_search "milk", ["Milk"], body_options: {min_score: 1}
   end
 
   def test_with_uneffective_min_score